Simplified MathJax
Rather than generating script tags, this now simply wraps suitable
delimiters around the maths, which MathJax will then find.  This is
essentially because I could not work out how to make the script tags
work with MatJax 3.

In fact I can't work out how the script tags ever worked: there doesn't
seem to be any mention of embedding maths this way even for MathJax 2:
all the focus is on using delimiters which MathJax looks for.

In both MathJax 2 and 3 this makes the output look nicer: using the
script approach seems to result in maths which, unless you apply
additional styling, is blue, while this approach results in maths which
is just ordinary.

This change

- only changes the output when using MathJax;
- makes the output simpler in that case;
- makes the formatted output better-looking I think;
- works with at least MathJax 2 as well as 3.